Move SingleTableInventoryCalculateParameter
